32-bit fixes

+++ comex/src-mpi3/reg_win.c
@@ -38,9 +38,9 @@ static reg_return_t seg_intersects(void *reg_addr, size_t reg_len,
 static reg_return_t seg_contains(void *reg_addr, size_t reg_len,
                                  void *oth_addr, size_t oth_len);
 static reg_return_t reg_entry_intersects(reg_entry_t *reg_entry,
-                                         void *buf, int len);
+                                         void *buf, size_t len);
 static reg_return_t reg_entry_contains(reg_entry_t *reg_entry,
-                                       void *buf, int len);
+                                       void *buf, size_t len);
 
 #define TEST_FOR_INTERSECTION 0
 #define TEST_FOR_CONTAINMENT 1
@@ -211,7 +211,7 @@ seg_contains(void *reg_addr, size_t reg_len, void *oth_addr, size_t oth_len)
  * @return RR_SUCCESS on success
  */
 STATIC reg_return_t
-reg_entry_intersects(reg_entry_t *reg_entry, void *buf, int len)
+reg_entry_intersects(reg_entry_t *reg_entry, void *buf, size_t len)
 {
 #if DEBUG
     printf("[%d] reg_entry_intersects(reg_entry=%p, buf=%p, len=%d)\n",
@@ -242,7 +242,7 @@ reg_entry_intersects(reg_entry_t *reg_entry, void *buf, int len)
  * @return RR_SUCCESS on success
  */
 STATIC reg_return_t
-reg_entry_contains(reg_entry_t *reg_entry, void *buf, int len)
+reg_entry_contains(reg_entry_t *reg_entry, void *buf, size_t len)
 {
 #if DEBUG
     printf("[%d] reg_entry_contains(reg_entry=%p, buf=%p, len=%d)\n",
@@ -402,7 +402,7 @@ reg_win_destroy()
  * @return the reg window entry, or NULL on failure
  */
 reg_entry_t*
-reg_win_find(int rank, void *buf, int len)
+reg_win_find(int rank, void *buf, size_t len)
 {
     reg_entry_t *entry = NULL;
     reg_entry_t *runner = NULL;
@@ -471,7 +471,7 @@ reg_win_find(int rank, void *buf, int len)
  * @return the reg window entry, or NULL on failure
  */
 reg_entry_t*
-reg_win_find_intersection(int rank, void *buf, int len)
+reg_win_find_intersection(int rank, void *buf, size_t len)
 {
     reg_entry_t *entry = NULL;
     reg_entry_t *runner = NULL;
@@ -520,7 +520,7 @@ reg_win_find_intersection(int rank, void *buf, int len)
  * @return return new entry in linked list
  */
 reg_entry_t*
-reg_win_insert(int rank, void *buf, int len, MPI_Win win, comex_igroup_t *group)
+reg_win_insert(int rank, void *buf, size_t len, MPI_Win win, comex_igroup_t *group)
 {
     reg_entry_t *node = NULL;
